chor: move trace functions to new file
