+++ src/ui/font-scaling.ts
@@ -3,9 +3,9 @@
 import { EventData, Observable, PropertyChangeData } from 'tns-core-modules/data/observable';
 import { isAndroid, isIOS } from 'tns-core-modules/platform';
 import { View } from 'tns-core-modules/ui/core/view';
+import { writeFontScaleTrace } from '../trace';
 import { FontScaleObservable } from '../utils/FontScaleObservable';
 import '../utils/global-events';
-import { writeFontScaleTrace } from '../utils/helpers';
 
 const fontScaleCssClasses = FontScaleObservable.VALID_FONT_SCALES.map(fontScaleToCssClass);
 const viewRefMap = new Set<WeakRef<View>>();
