+++ src/utils/helpers.ts
@@ -1,6 +1,5 @@
 /// <reference path="../ui/core/view.d.ts" />
 
-import * as trace from 'tns-core-modules/trace';
 import { Property } from 'tns-core-modules/ui/core/properties';
 import {
   AccessibilityBlurEventData,
@@ -9,6 +8,7 @@ import {
   booleanConverter,
   View,
 } from 'tns-core-modules/ui/core/view';
+import { isTraceEnabled, writeTrace } from '../trace';
 
 export function noop() {
   // ignore
@@ -43,7 +43,9 @@ export function enforceArray(val: string | string[]): string[] {
     return val.split(/[, ]/g).filter((v: string) => !!v);
   }
 
-  writeTrace(`enforceArray: val is of unsupported type: ${val} -> ${typeof val}`);
+  if (isTraceEnabled()) {
+    writeTrace(`enforceArray: val is of unsupported type: ${val} -> ${typeof val}`);
+  }
 
   return [];
 }
@@ -89,34 +91,6 @@ export function addBooleanPropertyToView<ViewClass extends View>(
   defaultValue?: boolean,
 ): Property<ViewClass, boolean> {
   return addPropertyToView(viewClass, name, defaultValue, booleanConverter);
-}
-
-export function isTraceEnabled() {
-  return trace.isEnabled();
-}
-/**
- * Write to NativeScript's trace.
- */
-export function writeTrace(message: string, type = trace.messageType.info, category = 'A11Y') {
-  if (isTraceEnabled()) {
-    trace.write(message, category, type);
-  }
-}
-
-export function writeFontScaleTrace(message: string, type = trace.messageType.info) {
-  writeTrace(message, type, 'A11Y-FontScale');
-}
-
-export function writeGlobalEventsTrace(message: string, type = trace.messageType.info) {
-  writeTrace(message, type, 'A11Y-GlobalEvents');
-}
-
-export function writeErrorTrace(message) {
-  writeTrace(message, trace.messageType.error);
-}
-
-export function writeWarnTrace(message) {
-  writeTrace(message, trace.messageType.warn);
 }
 
 /**
